# pkg/api/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "api.ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string"
                    }
                },
                "required": [
                    "error"
                ],
                "type": "object"
            },
...
            "engine.CreateNetworkParams": {
                "properties": {
                    "address": {
                        "type": "string",
                        "description": "Optional host IPv4 CIDR address for bridge mode. Does not configure DHCP or routing."
                    },
                    "device": {
                        "type": "string",
                        "description": "Optional existing native bridge device to borrow. Applied device ownership remains server-managed."
                    },
                    "members": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false,
                        "description": "Existing host interfaces to join a native bridge."
                    },
                    "mode": {
                        "enum": [
                            "bridge",
                            "switch",
                            "bridged",
                            "vmnet-bridged",
                            "user",
                            "vlan"
                        ],
                        "type": "string",
                        "description": "Network backend. Omission defaults to bridge."
                    },
                    "name": {
                        "minLength": 1,
                        "type": "string",
                        "description": "Network name. Reserved names user and vmnet-* cannot be used."
                    },
                    "parent": {
                        "type": "string",
                        "description": "Host parent interface for vlan mode."
                    },
                    "tag": {
                        "maximum": 4094,
                        "minimum": 0,
                        "type": "integer",
                        "description": "VLAN ID from 1 to 4094 when mode is vlan. Omit in other modes."
                    },
                    "uplink": {
                        "type": "string",
                        "description": "Host interface name from listInterfaces. Required for vmnet-bridged or bridged."
                    },
                    "vlans": {
                        "items": {
                            "$ref": "#/components/schemas/types.VLAN"
                        },
                        "type": "array",
                        "uniqueItems": false,
                        "description": "VLANs to create and join to a bridge using parent interface and tag."
                    }
                },
                "required": [
                    "name"
                ],
                "type": "object"
            },
...
            "host.Port": {
                "properties": {
                    "addresses": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "device": {
                        "type": "string"
                    },
                    "hardware_port": {
                        "type": "string"
                    },
                    "up": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "addresses",
                    "device",
                    "hardware_port",
                    "up"
                ],
                "type": "object"
            },
            "jobs.Job": {
                "properties": {
                    "action": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "error": {
                        "type": "string"
                    },
                    "finished_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "label": {
                        "type": "string"
                    },
                    "logs": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "result": {
                        "type": "string"
                    },
                    "started_at": {
                        "type": "string"
                    },
                    "state": {
                        "$ref": "#/components/schemas/jobs.State"
                    },
                    "target": {
                        "type": "string"
                    }
                },
                "required": [
                    "action",
                    "created_at",
                    "id",
                    "label",
                    "logs",
                    "state",
                    "target"
                ],
                "type": "object"
            },
...
            "jobs.State": {
                "enum": [
                    "pending",
                    "running",
                    "succeeded",
                    "failed"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "Pending",
                    "Running",
                    "Succeeded",
                    "Failed"
                ]
            },
...
            "types.NetworkManifest": {
                "properties": {
                    "address": {
                        "type": "string"
                    },
                    "applied_address": {
                        "type": "string"
                    },
                    "applied_members": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "applied_vlans": {
                        "items": {
                            "$ref": "#/components/schemas/types.VLAN"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "device": {
                        "type": "string"
                    },
                    "group": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "members": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "mode": {
                        "$ref": "#/components/schemas/types.NetworkMode"
                    },
                    "name": {
                        "type": "string"
                    },
                    "owned": {
                        "type": "boolean"
                    },
                    "parent": {
                        "type": "string"
                    },
                    "tag": {
                        "type": "integer"
                    },
                    "uplink": {
                        "type": "string"
                    },
                    "vlans": {
                        "items": {
                            "$ref": "#/components/schemas/types.VLAN"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "required": [
                    "id",
                    "mode",
                    "name"
                ],
                "type": "object"
            },
            "types.NetworkMode": {
                "enum": [
                    "user",
                    "vmnet-bridged",
                    "bridge",
                    "switch",
                    "bridged",
                    "vlan"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "NetworkUser",
                    "NetworkVmnetBridged",
                    "NetworkBridge",
                    "NetworkSwitch",
                    "NetworkBridged",
                    "NetworkVLAN"
                ]
            },
...
            "types.VLAN": {
                "properties": {
                    "borrowed": {
                        "type": "boolean"
                    },
                    "device": {
                        "type": "string"
                    },
                    "parent": {
                        "type": "string"
                    },
                    "tag": {
                        "maximum": 4094,
                        "minimum": 1,
                        "type": "integer"
                    }
                },
                "required": [
                    "parent",
                    "tag"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Bearer JWT obtained from POST /api/login.",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "VM lifecycle, networking, media and jobs API. Mutations returning 202 are asynchronous; inspect the returned job for completion.",
        "title": "Maco API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/interfaces": {
            "get": {
                "description": "List host network interfaces and addresses. Use their actual names for bridge members, VLAN parents, or vmnet uplinks.",
                "operationId": "listInterfaces",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/host.Port"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "listInterfaces",
                "tags": [
                    "interfaces"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": true,
                    "transport": "json"
                }
            }
        },
...
        "/api/jobs/{id}/wait": {
            "get": {
                "description": "Wait for one job ID until completion through a finite server-sent event stream. The done event returns the final job. Failed jobs are errors; connection loss does not cancel or prove failure of the underlying job. Inspect getJob or wait again with the same ID.",
                "operationId": "waitJob",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "text/event-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "text/event-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "text/event-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "503": {
                        "content": {
                            "text/event-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "waitJob",
                "tags": [
                    "jobs"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": true,
                    "stream": {
                        "event": "done",
                        "failureValue": "failed",
                        "stateField": "state"
                    },
                    "transport": "sse"
                }
            }
        },
...
        "/api/networks": {
            "get": {
                "description": "List configured networks and their IDs, modes, and applied topology. Use IDs from this response in VM interface requests. Applied fields describe engine-owned state, not authoring inputs.",
                "operationId": "listNetworks",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/types.NetworkManifest"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "listNetworks",
                "tags": [
                    "networks"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": true,
                    "transport": "json"
                }
            },
            "post": {
                "description": "Create a named network. Discover host interface names with listInterfaces. Mode bridge supports members, address (IPv4 CIDR), and vlans; vmnet-bridged requires uplink; vlan requires parent and tag (1 to 4094). Host address does not create DHCP or routing. Use the request schema rather than copying a stored network manifest. Requires an administrator account.",
                "operationId": "createNetwork",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/engine.CreateNetworkParams",
                                "summary": "body",
                                "description": "Request"
                            }
                        }
                    },
                    "description": "Request",
                    "required": true
                },
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "createNetwork",
                "tags": [
                    "networks"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": false,
                    "transport": "json",
                    "wait": {
                        "idField": "id",
                        "operationId": "waitJob",
                        "parameter": "id"
                    }
                }
            }
        },
        "/api/networks/{id}": {
            "delete": {
                "description": "Delete a configured network and clean up host resources owned by Maco. May disrupt attached VMs. Borrowed host interfaces are retained. Requires an administrator account.",
                "operationId": "destroyNetwork",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "destroyNetwork",
                "tags": [
                    "networks"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": false,
                    "transport": "json",
                    "wait": {
                        "idField": "id",
                        "operationId": "waitJob",
                        "parameter": "id"
                    }
                }
            },
            "put": {
                "description": "Replace the desired settings of an existing network using its ID from listNetworks. Read existing configuration first and supply the desired complete settings. This is not an arbitrary manifest replacement; applied bookkeeping is managed by Maco. Requires an administrator account.",
                "operationId": "updateNetwork",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/engine.CreateNetworkParams",
                                "summary": "body",
                                "description": "Request"
                            }
                        }
                    },
                    "description": "Request",
                    "required": true
                },
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "updateNetwork",
                "tags": [
                    "networks"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": false,
                    "transport": "json",
                    "wait": {
                        "idField": "id",
                        "operationId": "waitJob",
                        "parameter": "id"
                    }
                }
            }
        },
        "/api/networks/{id}/apply": {
            "post": {
                "description": "Reconcile the selected network configuration onto the host. Can affect host connectivity; inspect listNetworks and listInterfaces before applying. Requires an administrator account.",
                "operationId": "applyNetwork",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "applyNetwork",
                "tags": [
                    "networks"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": false,
                    "transport": "json",
                    "wait": {
                        "idField": "id",
                        "operationId": "waitJob",
                        "parameter": "id"
                    }
                }
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/"
        }
    ],
...
}